// frontend/mobile/src/types/navigation.ts
export type RootStackParamList = {
  Login: undefined;
  Signup: undefined;
  ForgotPassword: undefined;
  MainApp: undefined;
  HealthData: undefined;
  Profile: undefined;
  HeartRate: undefined;
  Activity: undefined;
  Sleep: undefined;
  Medicine: undefined;
  ModeSelection: undefined;
  Prediction: { mode: 'self' | 'other' };
};

export type TabParamList = {
  Home: undefined;
  Profile: undefined;
  Prediction: undefined;
  Misc: undefined;
}; 